#18 增加一下脱困方式

+++ src/sr/control/__init__.py
@@ -116,7 +116,7 @@ class GameController:
         """
         pass
 
-    def move(self, direction: str, press_time: int = 0):
+    def move(self, direction: str, press_time: float = 0):
         """
         往固定方向移动
         :param direction: 方向 wsad
